Part 2 of Day 20 problem checkpointing
This one is a bear.  The beary thing really is that for the first part of the problem I didn't really need to assemble the image to determine the requested answer.  But for this second part, I actually need to assemble.

Work in progress...  It almost works, but something is amiss...